[UPDATE] Updated the Collector and Collection class
Implemented methods, added comments and implemented a bidirectional relationship between Collection and Collector where the Collector is the leader when it comes to adding collections to the list.

Example:
Collection class:
void setCollector() is not publicly accessible, only package accessible, and it only sets the Collector for that collection.

Collector class:
public void addcollection() is publicly accessible and it adds a Collection to the list, while adding itself to that given Collection.